rename evm_call_tests to eth_call_tests
