+++ crates/evm/src/tests/queries/mod.rs
@@ -1,6 +1,6 @@
 mod basic_queries;
 mod estimate_gas_tests;
-mod evm_call_tests;
+mod eth_call_tests;
 mod log_tests;
 
 use std::str::FromStr;
